Added FileDataSource and functions to work with files. Added RecordsDataSource. Minor improvements.

+++ app/src/main/java/com/dimowner/audiorecorder/v2/data/PrefsV2.kt
@@ -29,7 +29,7 @@ interface PrefsV2 {
 
     var askToRenameAfterRecordingStopped: Boolean
 
-    var activeRecordId: Long
+    var activeRecordId: Int
 
     val recordCounter: Long
     fun incrementRecordCounter()
